[PL] Split UniformDirichletBC.cpp. No changes.
